requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.5
python-socketio==5.10.0
pywebview==5.0.5
python-engineio==4.8.0
simple-websocket==1.0.0
